Add checks to data preparation script
Checks that the bounding boxes are valid. Also now summarizes total
number of images written to the train and test sets. Some print
statements were removes to make the output more concise, and to
avoid burying warnings.

The labels are set for food detection, instead of can detection.
We should add a more universal label filter or allow the user to
specify custom filters for compatibility with non-food detection.