Add support for generation of datasets split by a label type
The argument --exclude/-e for generate_listdata.py can be used to
create separate training and testing sets: one set that excludes
images with objects with the given label and one that includes them.
Some of the images that include objects for this specific label may
still include objects for other labels.

This split is useful for testing the generalizability of a model.
I.e. a model trained on the excluded set can be tested on the
excluded set to show how well it performs on an unseen object type.